Fix for app crash when moving messages to the trash folder
#comment This is a simplistic patch that would reduce the probability of
app crashing when moving messages to the trash folder, in the absence of
a general solution that would fix concurrency. Additionally, it should
fix moving messages to spam, archive and custom folders.
Affected: moving messages to trash and other folders, swiping messages
to trash, from MailboxActivity and MessageDetailsActivity

Ticket: MAILAND-601