Defined DownloadEmbeddedAttachmentsWorker task as unique.
A bug has been reposter that in certain scenarios incosisted downlod attachment user feedback is diplayed. E.g. loading spinner next to the attachment is displaed forever, a toast with "Can't download attachment". It looks like this could be due to the fact that by touching the attachment user can start downlod multiple times and feedback reposted to UI by different instances is inconsisted. This fix will try to change it.

MAILAND-972