Observe saveLabel result to display outcome in LabelsManagerActivity
As this piece of logic is not covered by tests I decided to be
conservative and just shifted the logic from onLabelAddedEvent to the
observer. Part of the logic handles folders creation, but by looking at
the places where `LabelAddedEvent` is posted I don't see anything related
to folders, so it may be that we do not need to handle them here.

Minor code formatting

Move LabelRepository injection to ApplicationBindsModule

Rename displayLabelCreationOutcome() parameter to workInfo

Move saveLabel to being a suspend function

As we want to keep all the IO functions suspended for safety

Format PostLabelWorker to solve warnings

MAILAND-1023