Do not wrap PostLabelWorker logic in a runCatching lambda
Because `CoroutineWorker` already executes all the doWork code within a
try catch block, so a failure would automatically be returned in case
any exception is thrown as part of the execution

Minor code formatting

Fix test that checks LabelID param is passed to update a Label

As before ce48df1b76df6b51 we were executing the whole worker in a
runCatching block, the exception thrown when the `labelID` param was not
passed was handled by the business logic and an error was manually
returned.
As the worker already wraps any exception thrown by `doWork` in the
Result's exception, we're not wrapping it anymore and that caused this
test to fail both because the exception now results as thrown (because
the superclass CoroutineWorker.doWork logic is not executed by the test)
We're then catching the exception and verifying it manually

Remove unneeded labelResponse.label null check in PostLabelWorker

`label`is  initialised `by lazy` in
LabelFactory.createDBObjectFromServerObject method.
Such method maps the server response checking each field
for null values and throwing a RuntimeException if any, which means
  `Label` will never be null

MAILAND-1023

Remove duplicated equals and hashcode methods