Add 'CreateContactGroupWorker' Enqueuer to schedule this worker
It may look like this class and its Enqueuer are duplicates
of PostLabelWorker. We evaluated that this duplication is coincidental
and happens because of how concepts such as `Labels` and `Contact
Groups are represented in the backend.
Nonetheless we think having two separated use cases to perform these
actions is a clearer and more flexible solution for the client.

MAILAND-1020