Fix crash when applying label and archiving message
The logic to perform this operation is convoluted. The VM fires
`onLabelsChecked` callback registered in `MessageDetailsActivity`
passing 'null` for many of the params. Then the activity queries
the VM from the callback to get the data it needs.

This crash started happening on [this commit](f6a1623384) as we
converted `MessageDetailsActivity` to kotlin. The last param of the
callback `onLabelsChecked` became a non-null value, which caused the crash
as null was passed.

The crash doesn't happen when applying a label without selecting `also
archive` because a different method is used to perform that operation.

MAILAND-1005