Fix incorrect toast showing when label is created
As part of MAILAND-1023 the way LabelsManagerActivity reacts to labels
creation performed by the worker was changed. The new implementation
didn't account of all the states of the worker that aren't 'SUCCESS' or
'FAILURE', causing the "Label not added due to invalid name." toast to
be shown before the creation process would complete.

The issue was fixed by adding a check to ensure we only display the
label creation outcome when the state of the worker is final

MAILAND-1039