Remove unused name field in ContactGroupDetailsActivity
This field was used by onPrepareOptionsMenu to set the title. It was
only needed as we had a listener on appBar offset changes to
invalidate the menu all the time. Since we removed such listener in
commit `ed7b4f75c0aa01e` and the setTitle function is called directly in
both places where we are setting the field, it's not needed anymore

MAILAND-750