Replace local contact emails with server's ones
As for the previous logic, after saving a contact on our API we delete
all the local emails and store only the new ones. This logic should be
revised to have a more consistent offline mode.
Using "Gson" to perform Json deserialization (from worker's data to list
of server email) on the use case is not optimal, as this kind of op
shouldn't belong to use case. Nevertheless, the scope increase of
extracting also a mapper for it didn't seem worth right now so I went
for this tradeoff.

As fot the previous commit, force unwrap was used to get the
`serverEmails` argument as we trust the worker will always return it
when succeeding. In case of any exceptional state, I evaluated a crash
to be more consistent then removing all of the contact's emails locally

MAILAND-1021