EditContactDetailsViewModel uses CreateContact UseCase to create user
Over relying on repository directly to schedule a job.
Unfortunately there was a decision to add no unit tests to
verify this new dependencu, as after investing some time in trying to
get this class in a test harness I figured that'd take too long
as it requires some changes to how this VM operates first.

MAILAND-1021

Observe createContact result and show outcome

on EditContactDetailsActivity. This might not be enough as originally
the outcome was displayed through the eventBus on the
ContactListFragment as well after this activity was closed

MAILAND-1021

Worker returns failure when contact details are not returned and no
specific error code is returned from backend

This avoids the CreateContact use case to crash as when success is
returned we expect some output data to be returned by the worker (which
in case of missing contact details in the response is not)

MAILAND-1021