Finish EditContactDetailsActivity only once op result was received
So that we ensure that some feedback about the contact creation is shown
to the user before going back to the contacts list.
In case of no network connectivity, the CreateContact use case will not
return currently.
We will add a functionality to return a message "Contact saved, will be
synced when connection is available" as the previous implementation was
doing so that we ensure the user doesn't get blocked.

MAILAND-1021