Move "contact saved" toast to EditContactDetailsActivity
`792e1ffe00b07` changed the behavior of this activity to not be
finished immediately after firing a contact creation / update. This
introduced the side effect that the `update` job result (which is still
returned through EventBus) is not received in EditContactDetailsActivity
instead of ContactDetailsActivity.

As no other job should be relying on ContactDetailsActivity to show the
success result, I removed the toast from there

MAILAND-1143

Add rontact-saved-offline message as a default one. This will be used
from updateJob when there's no network (unitl that job gets migrated
too)