Fail sending message when attachments upload failed
Currently we're ignoring the result of the attachment upload API call,
which might cause messages to be sent without attachments.
This will block the sending of the message, show a notification to
communicate it to the user and remove the message from the list of
messages pending to be sent to avoid the message to be forever pending
for send in draft.
User is required to manually send this message again from drafts

MAILAND-1139