UploadAttachments skips uploading if file does not exist
As we need to perform File operations and we do not want to write to
disk, we need some abstraction on the File creation to be able to mock
`.exists` call. I choose to wrap this file operation in the attachment
model and mock them in the use case over creating a new class to handle
files in order to keep the scope of this task from expanding too much.

MAILAND-1139