CreateDraftWorker reads parent message from messageDetailsDB
and sets action on request when parent Id is given

Here we are dropping some functionality by reading parent message only
from messageDetailsRepository instead of using searchDB when
`isTransient` is true. In this instance, isTransient seems to only be
true when all the following conditions verify:
- Search for a message
- Tap on a message _that you previously sent_ from the search result
- Perform an action on the message (reply, reply_all, forward)
In this case is transient is true and the parent message is read from
searchDB. The same message is also available in messageDB at the same
point in time, with just the following differencies:
SEARCHDB | MESSAGEDETAILSDB
- isInline (false | true)
- location (2 | 3)
- spamScore (0 | -1)
- accessTime (1607507146022 | 0)
which doesn't seem to create a difference in functionality. Will come
back to this decision when knowing more on `isTransient`

MAILAND-1018