SaveDraft removes pending upload when a failure occours
Either in the save draft API call itself on in the UploadAttachments use
case.
The responsibility to `insert` and `delete` pending upload would fit
better into `UploadAttachments` itself. I decided not to move it yet as
that'd change behaviour in case the user takes actions on the draft
while it is still uploading (as pendingUpload wouldn't be inserted yet)

MAILAND-1281