Allow attachments to be added when draft was not created on API yet
This was also the behavior before the saveDraft refactoring took place,
as despite the logic in `AddAttachmentsActivity` seems to be in place to
avoid it, the `PostMessageServiceFactory` was posting a
`DraftCreatedEvent` which when received in `AddAttachmentsActivity`
would set mDraftCreated field to true (even if the draft was just local
like in the case of no connectivity).
The mentioned logic that received the DraftCreatedEvent was deleted by
commit 779a328 as part of this refactoring.
This same behavior is achieved by not considering `isLocalMessageId` to
comptute EXTRA_DRAFT_CREATED boolean.

Except for an edge case for which the user clicks the "add attachments"
menu immediately after opening the composer, this condition will always
be true. Tracked in a task to be completely removed after this refactoring

MAILAND-1281