CreateDraftWorker updates pendingUpload with ID of created draft
This is needed to ensure attachments upload is being done also when the
app is killed immediately after creating a draft with no connectivity.

- Local Message attachments are preserved after creating a draft

Tech details:
As the upload of attachments happens in `saveDraft` use case when the
worker's API request finishes, in case of app being killed it won't be
called (as only the worker will get executed) causing the
"PendingUpload" stored at the beginning of the process to be wrong as it
contains the id of the local message instead of the newly created remote one.