Allow drafts to be saved for messages that are pending for send
The piece of logic blocking "save draft" was introduced to replicate
the logic existing in `CreateAndPostDraftJob` (deleted by 4ca172ca5895f6f5).
The logic was in place because the sending process would perform
its own direct API call to save the draft, thus the job was not needed.

It is not the case anymore, as the sending process now goes through the
same `saveDraft` use case to save a draft, thus we want it to be able to save.

MAILAND-914