Inject addressCrypto into PackageFactory to make the class injectable
Since this change to the class couldn't be performed through fully
automated refactorings as the two previous changes, we added a test to
cover it.
Right now we compromised of making it an android test as there are some
dependencies within Crypto's `SessionKey` object which could not be
broken effectively without doing too many changes. The added test isn't
too detailed as its main goal is to ensure the instance of crypto being
used (now created through `AddressCrypto.Factory`) is working well

MAILAND-914