Do not update draftId with value returned by AddAttachmentsActivity
Changes:
- Do not update viewmodel's draftId with the one returned by AddAttachmentsActivity
- Observe outcome of draft update (not only creation) from VM (as with
  auto save it's now possible this flow is executed while the composer
  is still open)

This solves a regression found during testing of build 756, for which
following the steps outlined in MAILAND-1300 an invalid, duplicated copy
of the draft would be created locally. The attachments added to the
draft would get lost.

The cause was this bit of logic in ComposeMessageActivity that "updated"
the VM's `draftId` with the value that was initially passed to
AddAttachmentsActivity. Since while the user is on `Attachments` screen
the update draft flow gets triggered by the auto save functionality, if
the user statys in the attachments screen long enough for the update to
complete the ID that we have on VM will be the updated one, which
differs from the local one that Attachments screen will return.

This wasn't happening before the refactor introduced by MAILAND-1281 as
the AddAttachmentsActivity would receive an `Event` through the bus with
the updated ID of the draft (which was then returned)

MAILAND-1300