Send Message worker shows error to the user and fails when API response is failure
This will happen when a response was successfully retrieved but the body
code is not SUCCESS nor VERIFICATION_NEEDED. It's a generic failure
which is then then notified it to the user through a push notification.
This message is removed from the 'pending for sending' list and the work
fails.

MAILAND-914