Log warning when decryptedBody passed to saveDraft is null
This case was not handled beforehand, and we took a call not to handle
it right now as it just depends on the call to this object being done
with the right object. This log is in place to grant observability over
this and ensure it's not happening in any other instance.

MAILAND-914