Handle failures while building Send Message API request
In some cases, PackageFactory.generatePackages method will throw
exceptions. One case might be (as example) uploading of attachments
being ongoing in parallel with sending the message. Since these are
cases we can recover from, the handling have been implemented in the
form of retry or fail depending on maxRetries.

MAILAND-914