Delete MessageSentEvent and its listeners as it's not used anymore
It was replaced by the new `SendMessage` implementation which uses
`userNotifier` to display the outcome of the sending process instead of
the bus.
In one case, the event was causing `AlarmReceiver` to be executed
immediately performing a refresh of the folder the user is in.
As it turned out from the tests done, the interested message is updated
also through a db observer, making that refresh superflous.

MAILAND-914