Save parent attachments which were already uploaded when draft is created
This is needed to avoid cases where some parent attachments (namely,
inline images) would cause an error  when sending a reply / forwarded
message.
It wasn't the case before the refactoring to "send" component  as the attachments
returned by draft creation were immediately used to perform the upload without
storing to DB as an intermediate step.

Since with the separated components that were introduced we now rely on
`CreateDraftWorker` to save the draft (with attachments and all the updated fields)
to DB and return its ID for `UploadAttachments` to perform its work, this piece of logic is
needed to ensure both attachments that were uploaded with the create
draft request(eg.inline images) and attachments that weren't uploaded
(any attachment the user added) are preserved.

MAILAND-914