Avoid corrupted state when killing app immediately after sending a message
This issue was happening on messages with attachments, and was caused by
the fact that killing the app once upload attachment did start was
causing "PendingUpload" to be left as true while no actual upload was
ongoing anymore.

As a consequence, this caused each subsequent retry to be denied as
"Pending upload" was found true and being interpreted as an upload being
in progress.

Solution consisted in moving the UploadAttachments logic to an android
Worker (so that it proceed in the backgrpind and handle cancellation properly)
and also remove "PendingUpload" when the worker finds it true, as that
means that the previous worker was cancelled (likely replaced by this one)

MAILAND-914