Upload Public key to the message only when sending it
This solves an issue for which the public key was uploaded as part of
saving a draft, each time it was saved.

This issue was introduced with commit 6319af6bef33122b94266d, which began re-using
the UploadAttachments use case logic without considering that while such logic
does indeed upload attachments, it was extracted from the sending process
(MAILAND-1139) and it contains some op (namely, upload of oublic key)
that are expected to be done only while sending and not when saving a draft.

MAILAND-1511