Fix crash when forwarding messages with a big body
The crash was happening for messages with a body bigger then 10240 bytes
as that's the limit of data that can be serialized as input in a Worker.
The solution consisted in encrypting the message body and storing it to
DB instead of passing the decrypted body to to the worker.
Consequentially, saveDraft changed in order to not rely on the existence
of a decryptedBody when being called during the sending process.

MAILAND-914