Do no explicitly handle "Upload in progress" state in UploadAttachments
As this is now handled implicitly through the Cancellation of the
unique worker when a new upload task is scheduled (for example, because
the app was killed or because the user saved the draft manually and then
re-opened it and sent it before the initial upload finished)

As part of the above, also change AttachmentsRepository to only catch
and handle IOException and rethrow every other exception (eg.
JobCancellationException). This enables any ongoing (but not finished)
API call to be cancelled avoiding duplicated attachments.
The exception thrown inside the UploadAttachments worker will cause the
worker to retry automatically.

MAILAND-914