Solve issues for which removed parent attachments were sent of Fw
THis issue was happening because removing parent attachments in the app (for a
message beign forwarded) was not being reflected in the backend.

The underlying cause is that performing a `copy` of the draft's creation
response, together with setting local attachments on the created draft,
was causing the loss of the reference to any attachments that were
created through the create draft request (such as parent attachments).

Consequentially, we did not have the correct ID of the attachment when
trying to perform the deletion, causing the response to be handled
silently (as it's the same case of the attachment not existing remotely) and
the attachments to be sent with the message.

MAILAND-1555