Attachments can be removed also before draft creation finishes on FW
This solves a regression for which if a user access the
attachments screen and removes any (parent) attachments before the
initial draft creation finished, those attachments would not be removed
and sent with the message instead.

The underlying cause for this regression is that until the draft is
created, attachments do have a "localId" which is not the same that
backend generates at draft's creation time.
For this reason, the removal fails silently as backend doesn't find any
attachments with such ids.

The solution consists in making AddAttachmentsActivity listen for
changes to the Message object in the DB, displaying the attachments only
after the draft creation finished (or immediately, in case the user is
offline)

MAILAND-1555