Cancel pending create draft requests before sending
This is needed to replicate the existing logic of not saving a draft
when sending is ongoing, (performed though the check for pendingForSend
in saveDraft use case) also when the message is sent while offline.

This prevents a race condition between saving the draft and sending
which might cause the message to be sent with attachments that were
instead removed from the offline message.

Moving the pendingForSend check to CreateDraftWorker was also evaluated,
but eventually this solution was preferred as that check was removed in
the send refactor branch which is currently under QA (so this solution
will apply to those changes too)

MAILAND-1555