Parent message data is added to draft API request only at first draft creation
Data as parentId and parent's attachments key packets need to be treated
only at the initial draft creation. For subsequent draft updates, all of
this data will be treated through the "normal" logic for the message.

eg. In the case of parent attachments when forwarding a message, the
attachments will be added to the draft creation request and become
message attachments equally as every other attachment the user might
add. On subsequent draft updates, they will be treated as such.

MAILAND-914