Solve issue where active VPN plan is lost when subscribing a Mail paid plan
This issue was caused by the API call to `createUpdateSubscription` API
being performed with only the `planId` of the Mail plan the user just
chose, omitting any existing plans the user already has as described
 by the `fetchSubscription` API response.

Technically speaking, said existing plans were being omitted because an
exception was thrown when `CreateSubscription` class was trying to add
the planIds returned by the API to the `planIds` list.

The core of the issue lies in the callers of `CreateSubscription`, which
as `planIds` parameter pass lists that are not mutable.
For example, the call from UpsellingActivity is performed passing
`Collections.singletonList(mSelectedPlanId)` which is fine at compile
time but throws exception at runtime as it's not mutable.

The solution was to change planIds to a mutableList directly in
`CreateSubscription` before performing any operations on it.

Other changes:
- Equals was added to CreateSubscriptionBody POJO for unit testing purposes

MAILAND-1640