Re-apply "Save parent attachments which were already uploaded when draft is created"
This reverts commit 4a9e63810cbecca72ead972a5c4e502b14cb66da.

This was removed to replicate the fix introduced for MAILAND-1555 in
release branch. Since the logic that handles drafts / attachments /
sending was changed in develop due to refactoring of the send
functionality (MAILAND-914) this piece of logic that preserves local
attachments doesn't cause the issue detailed in MAILAND-1555 anymore,
and it actually prevents other issues with attachments being out of
sync compared to what the user sees locally.

In particular, this handling local attachments prevents an issue for which
attachments would not be sent when adding them to a FW message
that had already attachments (parent attachments)

MAILAND-1665