Do not explicitly cancel ongoing draft creation when sending message
This was introduced by 0254ae69f1 to solve a race condition between the
ongoing save draft and the sending operation, which was causing
attachments being sent to differ from attachments the user was seeing.

This erroneous behaviour was happening as the logic to create a draft in
"SaveDraft" use case was differing from the one in "PostMessageJob".
Since SendMessageWorker was introduced the logic for saving draft is
exactly the same (SaveDraft use case) at any time (be it saving a draft
or sending). The case of sending a message while a draft being saved
will be handled implicitly from the SendMessageWorker, which  by
re-scheduling the SaveDraft use case will cancel the current job.

- Decrease detekt maxIssues

MAILAND-1665