SaveDraft use case returns Result instead of flow of Result
This makes collection simpler, as we always only care about one value in
this case.
It also avoids some cases where the collection of the "saveDraft" flow
would block execution of a method and not allow code after the call to
save draft to be executed (eg. usage in ComposeMessageVM was behaving
like this)

MAILAND-1685