Fix error sending message when sender was changed multiple times
This error was triggered by some of the attachments being changed more
than twice on a message for which some of the attachments were already
uploaded. The underlying cause, was that "oldSenderId" field on
ComposeMessageVM was not updated when being changed the second time,
causing the re-encryption of the attachment to fail (as we couldn't
decrypt it with the key of the wrong 'oldSender') and consequentially
the send operation to fail.

The fix consisted in always updating the old sender field with the
value of the current sender once the draft was updated. To ensure this
happens timely, we also trigger saving a draft when the sender changes.

MAILAND-1685