MessagesRecyclerViewAdapter shows MailboxUiItem objects
This is the first step needed to enable Mailbox to display both Messages
and Conversations. It just consist in wrapping the data that the
adapter needs to show an item into an object that can be used to show
any of the two kinds (message / conversation).

Some behavior was purposefully not implemented by this commit:
- Definition of sender name to be shown
- Definition of 'isBeingSent'
- Definition of 'uploadingAttachments'
This decision was due to the fact that it wasn't worth to replicate this
logic in the activity, as we have the data we need to compute it in the
VM, where the message -> mailboxItem mapping will be pushed in the next
commit

Changes:
- MessagesRecyclerViewAdapter and MessagesListItemView get a List of MailboxUiItem
- `onItemClick` and `checkedMessages` return MailboxUiItems
- Simplified `SimpleMessage` class removing unused fields
- Mailbox and Search activities map messages to mailboxItems when
  calling the adapter

MAILAND-1646