Add a v4 bottom action view in message details
The view contains actions for reply/reply all, marking a message as unread, moving a message to trash and an action for opening an action sheet that will be implemented in another ticket. The logic for the working actions is the same from before adding the v4 look to the MessageDetailsActivity.

MAILAND-1712