Issue #67 - Disable swiping actions - feature added
At this stage, the feature works as wanted and expected. The only downside is that the app needs to be closed and reopened in order to the change applied in the MailboxActivity. But I think it is good enough for the users wanting to disable the swiping gestures.
Affected: EditSettingsItemActivity.kt - MailboxActivity.kt - SettingsEnum.kt - User.kt - Constants.kt - edit_settings_structure.json - strings.xml