Remove "isTransient" flag from everywhere
Such flag was used to differentiate between messages that were coming
from "Search" (being the transient ones) and those that weren't.

This conditional being performed in many places was adding a lot of
noise and not really bringing any value, as the same behaviour can be
achieved by not storing messages from search in the first place.