Pull definition of "locationId" from GetConversations to the caller
This is because (as for the previous commit) the VM (caller of
GetConversations use case) also needs to know the locationId (be
it a "standard" location or a custom label) that the mailbox items are
being requested from. Not doing it also in the use case avoids
duplication of the logic which decides whether the location is a
standard or custom one and assignes the Id accordingly.

MAILAND-1437