Avoid duplicated refresh of Mailbox when switching account
`switchToMailboxLocation` is changing the value of `mailboxLocationMain`
liveData, which triggers a call to `observeMailboxItemsByLocation`,
making the other "explicit" call to `observeMailboxItemsByLocation`
a duplicate. Also, since no clear reason to "delay" the operation was
visible so far, the Handler.postDelayed was dropped.

As the "refreshMessages" param also defines whether the
"messageDetailsRepo" dependencies in the VM are reloaded, it needs to be
set to true when switching account (otherwise the old account items
might be wrongly displayed in the mailbox).
tHis will be done in a subsequent commit (d8660bf8b1c9cf9)

MAILAND-1822