Fix error for which MessageRepository would not save or return messageBody
when message was meant to be saved to DB.

`saveMessage` checks if the body of a message exceeds some limit and
saves it to a file when it does. In this case, the file path is set
as `messageBody` for the message that then is saved in the DB.

The issue is that when messageBody was not too big, null was being set
as messageBody and saved to the DB. Such null messageBody was also
retruned from methods such as `getMessage` because of the
Usage of `apply` on the `saveMessage` input param (which caused
the message passed by the caller to change and have a null `messageBody`.

MAILAND-1767