Prevent attachments view from randomly not being displayed
This issue started happening after 648f4710e56a4797360e1. The bit that
was preventing this was the flag `isDecrypted` which was ensuring the
view was only being called once with a given message.
Mutiple calls to the view cause the issue mentioned above, for which
opening the same message multiple times causes the attachments view to
being shown / hidden randomly.
No issue was found in the attachments data itself (which is always
valid). The assumption is that the issue lies in the adapter's logic
which resets / invalidates the attachments view when showing the
message. This will be tracked in the backlog as tech debt.

- Do not emit the same data twice (distinctUntilChanged)

MAILAND-1767