Display attachments for each message in a conversation
By pulling the creation of the "attachments adapter" and the displaying
logic into the adapter (so that it gets executed for each item).
Twp pieces of functionality to "set is pgp encrypted" and hide the
"download attachments" spinner were dropped, as migrating them implied a
relevant effort and the existing plan to change attachment presentation
will invalidate such features anyways.

- Manually changed AttachmentsView to MessageDetailsAttachmentView in
  layout_message_details_web_view while rebasing over
  609fdbb596a37c2381d03ec3fbf08b5b3d53aed0 changes

MAILAND-1770