Adapt 'load embedded images' functionality to work for conversations too
Beforehand, this functionality was working by injecting a string (the
html body with the images) into the adapter which would replace any
content of the webview with the new string. As with conversations we can
have multiple webviews, this didn't work anymore.

The new approach leverages the components currently in place to avoid
touching on a too wide scope: `MessageRenderer` is still in charge of
creating a message body with inline images. Such body is now being
returned with the messageId to which the body refers too. It's observed
in the MessageDetailsViewModel.init method and when fired it "hot swaps"
the `decryptedHtml` on the correct message and emit the
convresationUiModel again for the adapter to refresh the displayed data.

Summary of tech changes:
- Changed MessageRender to return an object containing both the
"rendered message body" and the messageId for which such body was rendered
- Removed `distinctUntilChanged` from _decryptedMessageLiveData as it
prevented embedded images from being emitted (as Message.decryptedHtml
field is not considered for computing equality)

- Rename `embeddedImagesAttachments` field following naming conventions
- `prepareEmbeddedImages` method receives message for which images should be prepared
- Message.decryptedHTML field can be set from outside of the model
- Remove redundant downloadEmbeddedImagesResult live data

MAILAND-1770