Display attachments when body is loaded for a message in a conversation
This solved an issue for which attachments where not shown when opening
a message which is not the latest (in a conversation) and which was never
opened before.
In such case, the message would be returned with an empty
`message.attachments` list even tho it has attachments, as they are only
correctly set on the message model when fetching the message from remote.

MAILAND-1931