Change lastMessage() method to actually return the last message
As ordering is done in a different place this method wasn't always
returning the last message, causing some issues hard to understand (eg.
embedded images randomply not loading for some converstions...)

Even tho this method will have to be removed soon, this change makes it
consistent for the moment being.

MAILAND-1931