Solve issue causing embedded images to not be loaded
when opening a msg the first time.

The issue was caused by a mistake in the instance of "message" which was
being passed to the `handleEmbeddedImagesLoading` method. Passing the
`fetchedMessage` ensures we have the correct attachments which are used
to infer whether embedded images should be displayed or not.

Other minor changes:
- Rename `handleEmbeddedImagesLoading` method to make the fact that this
  method actually loads images a little bit more explicit
- Change the `is message read` check to be performed on the
  fetchedMessage to ensure data used to decide is coming from out source
  of truth (to avoid another case as the one fixed in this commit)

MAILAND-1931