Solve crash when scrolling across conversation messages while images are loading
If the user clicks the 'load embedded images' button for one message and
then quickly scrolls / opens another message, by the time the images are
loaded and the adapter receives the callback the item won't be visible
anymore and thus it won't be found.

Treating the item as optional solves this. Updating the item through
`allItems` list proved not viable as it doesn't provide a consistent
access to the right items with the current implementation of the adapter.

MAILAND-1931