New location resolution logic in MessageFactory.
Message floder location is not directly returned by the server when we fetch the messages. There is a dedicated logic in MessageFactory that tries to determinate it from locationIDs array returnse with the message from the server. It looks like this logic didn't work well as messages were dissapearing from the folders after message fetch from the server. It has been updated according to the information from https://confluence.protontech.ch/display/MAIL/Message.

MAILAND-1558