Added messagesDao dependency to MessageLocationResolver.
This is needed to resolve our location from label id as when we get messages from the backend we get only locationIDs, including ids of a label, however we do not know if this label is really just a label or a folder, to know that we have to crosscheck the label id received from labels api.
However this currently does not fully work due to our current setup, we are unable to inject mesasges dao or anything else into MessageResponse object, therefore currenlty any labels location will be resolved just as a Label.
Hopefully move to core network would make this possible.

MAILAND-1558