Error state from mailbox in case of no messages has been added.
We didn't have an error state screen defined for mailbox in the situation when there was no messages fetched and an error has occured.
This has been added to prevent displaying just an empty screen in this situation.
If the mailbox recycler view has already any messages dispalyed, just a toast with an error message  will be displayed as before.