Added new pull to refresh logic in messages list.
The currently used pull to refresh logic based on MessageService seems to be not enough to fix some edge cases with no connectivity, therefore it has been now integrated with the main messages fetching flow and MessageRepositor. The main change is that now on refresh we will restart the fetching flow.
Also if network connection fails we will try to emit the current data from the db first.