The "server not reachable" banner was blinking and showing up multiple times. Because when some specific exceptions were received from the server we were assuming that server is not reachable, right away.
This commit introduces a new connectivity state, that is emitted in this situations, when we a call to the api resulted with an exception. In that case we are not sure if the server is reachable, so the state is called PING_NEEDED and after that we ping the server and only show the banner once the ping returns a result.

MAILAND-1807