Extract logic to break message body into parts to its own class
The main goal of this refactor is to add some (very basic) unit tests on
the logic that splits a message body in different parts and to return an
object holding the splitted message bits over a Pair.

Note that for a time tradeoff  `MessageBodyParser` was left using "Jsoup"
in a static way to perform the parsing of the html.

MAILAND-1533