Action Sheet applies actions to a specific message when summoned for a message into a conversation
Adding the "in message actions" bar at the bottom of each message in a
conversation added this new case: `conversationMode` will be enabled but we
will not want to call a conversation's endpoint to perform the given
action, as it was triggered for one specific message.

`MessageActionSheet` "originator" arg was used to infer the origin of
the call and to change endpoints and dismissing behavior accordingly.

MAILAND-1533