Solve issue for which "moving message" feature was not working
This issue got introduced by `a389c5965a71e336cc75b9` which changed the
internal logic of `MoveMessageToFolder` use case to check on the string
value of `MessageLocationType.messageLocationTypeValue` (which is the ID
of the location) instead of the String value of the enum's name.

Specifically, the issue appeared as after the mentioned change, clients
of `MoveMessageToFolder` were not changed to pass the data differently
(ID instead of name). Compilation didn't catch this as both values were
required as Strings.

As an initial "safety measure", this commit disallows the usage of
`.toString()` on any MessageLocationType enum values.

MAILAND-1533