Fix crash in Message Location Resolver logic under specific message conditions
Because of a mistake in the list from where we were reading the label,
under some specific conditions (eg. labels list containing a custom
folder's ID at one of the indexes checked by the `shortLabels` loop) the
application was crashing when performing a "move message" action through
the ActionSheet.

The crash was happening on the `item.toInt()` cast as the item being
casted was a custom folder instead of a standard location.

MAILAND-1533