Action Sheet's Move to folder functionality now works when moving a message inside a conversation's detail
Similarly to what explained in commit `ff2db8a082bd6b2282`, this commit
adds handling for the new case of an actions being applied to a message
into a conversation.

To make it clearer, both MessageActionSheet and LabelsActionSheet now
take a `ActionSheetTarget` (enum) instance as a param: such enum helps
defining what entity the action will be applied to depending on the
screen that summoned the actionsSheet and the value of "conversationMode".

There is some duplication between the two action sheets which was
knowingly left behing: it can be extracted to a composable component
at any later time.

MAILAND-1533